passes module integrated
