+++ qbraid/transpiler/conversions/cirq/cirq_to_qasm2.py
@@ -15,10 +15,50 @@ Module for conversions between Cirq Circuits and QASM strings
 from typing import Optional
 
 import cirq
-from cirq import ops
+from cirq import ops, value
 
 from qbraid._version import __version__ as qbraid_version
-from qbraid.transforms.cirq import map_zpow_and_unroll
+
+
+@value.value_equality
+class ZPowGate(cirq.ZPowGate):
+    """A single qubit gate for rotations around the
+    Z axis of the Bloch sphere.
+    """
+
+    def num_qubits(self) -> int:
+        """The number of qubits this gate acts on."""
+        return 1
+
+    def _qasm_(self, args: "cirq.QasmArgs", qubits: tuple["cirq.Qid", ...]) -> Optional[str]:
+        args.validate_version("2.0")
+        if self._global_shift == 0:
+            if self._exponent == 0.25:
+                return args.format("t {0};\n", qubits[0])
+            if self._exponent == -0.25:
+                return args.format("tdg {0};\n", qubits[0])
+            if self._exponent == 0.5:
+                return args.format("s {0};\n", qubits[0])
+            if self._exponent == -0.5:
+                return args.format("sdg {0};\n", qubits[0])
+            if self._exponent == 1:
+                return args.format("z {0};\n", qubits[0])
+            return args.format("p({0:half_turns}) {1};\n", self._exponent, qubits[0])
+        return args.format("rz({0:half_turns}) {1};\n", self._exponent, qubits[0])
+
+
+def map_zpow_and_unroll(circuit: cirq.Circuit) -> cirq.Circuit:
+    """Map ZPowGate to RZ and unroll circuit"""
+
+    def _map_zpow(op: cirq.Operation, _: int) -> cirq.OP_TREE:
+        if isinstance(op.gate, cirq.ZPowGate):
+            yield ZPowGate(exponent=op.gate.exponent, global_shift=op.gate.global_shift)(
+                op.qubits[0]
+            )
+        else:
+            yield op
+
+    return cirq.map_operations_and_unroll(circuit, _map_zpow)
 
 
 def _to_qasm_output(
