+++ qbraid/transpiler/conversions/qasm3/qasm3_to_braket.py
@@ -16,14 +16,41 @@ from typing import TYPE_CHECKING
 
 from qbraid_core._import import LazyLoader
 
+from qbraid.passes.qasm3.compat import (
+    convert_qasm_pi_to_decimal,
+    remove_stdgates_include,
+    replace_gate_name,
+)
 from qbraid.programs import QasmError
-from qbraid.transforms.qasm3.compat import transform_notation_to_external
 
 braket_circuits = LazyLoader("braket_circuits", globals(), "braket.circuits")
 braket_openqasm = LazyLoader("braket_openqasm", globals(), "braket.ir.openqasm")
 
 if TYPE_CHECKING:
     import braket.circuits
+
+
+def transform_notation(qasm3: str) -> str:
+    """
+    Process an OpenQASM 3 program that was generated by
+    an external tool to make it compatible with Amazon Braket.
+
+    """
+    replacements = {
+        "cx": "cnot",
+        "sdg": "si",
+        "tdg": "ti",
+        "sx": "v",
+        "sxdg": "vi",
+        "p": "phaseshift",
+        "cp": "cphaseshift",
+    }
+
+    qasm3 = remove_stdgates_include(qasm3)
+    for old, new in replacements.items():
+        qasm3 = replace_gate_name(qasm3, old, new)
+    qasm3 = convert_qasm_pi_to_decimal(qasm3)
+    return qasm3
 
 
 def qasm3_to_braket(qasm3_str: str) -> "braket.circuits.Circuit":
@@ -39,7 +66,7 @@ def qasm3_to_braket(qasm3_str: str) -> "braket.circuits.Circuit":
         CircuitConversionError: If qasm to braket conversion fails
 
     """
-    qasm3_str = transform_notation_to_external(qasm3_str)
+    qasm3_str = transform_notation(qasm3_str)
 
     try:
         program = braket_openqasm.Program(source=qasm3_str)
