+++ qbraid/transpiler/conversions/qasm3/qasm3_to_qiskit.py
@@ -16,13 +16,38 @@ from typing import TYPE_CHECKING
 
 from qbraid_core._import import LazyLoader
 
-from qbraid.transforms.qasm3.compat import transform_notation_from_external
+from qbraid.passes.qasm3.compat import add_stdgates_include, insert_gate_def, replace_gate_name
 
 qiskit_qasm3 = LazyLoader("qiskit_qasm3", globals(), "qiskit.qasm3")
 
 
 if TYPE_CHECKING:
     import qiskit as qiskit_
+
+
+def transform_notation(qasm3: str) -> str:
+    """
+    Process an OpenQASM 3 program that was generated by
+    an external tool to make it compatible with Qiskit.
+
+    """
+    replacements = {
+        "cnot": "cx",
+        "si": "sdg",
+        "ti": "tdg",
+        "v": "sx",
+        "vi": "sxdg",
+        "phaseshift": "p",
+        "cphaseshift": "cp",
+    }
+
+    for old, new in replacements.items():
+        qasm3 = replace_gate_name(qasm3, old, new)
+    qasm3 = add_stdgates_include(qasm3)
+    qasm3 = insert_gate_def(qasm3, "iswap")
+    qasm3 = insert_gate_def(qasm3, "sxdg")
+
+    return qasm3
 
 
 def qasm3_to_qiskit(qasm3: str) -> "qiskit_.QuantumCircuit":
@@ -39,6 +64,6 @@ def qasm3_to_qiskit(qasm3: str) -> "qiskit_.QuantumCircuit":
     except qiskit_qasm3.QASM3ImporterError:
         pass
 
-    qasm3 = transform_notation_from_external(qasm3)
+    qasm3 = transform_notation(qasm3)
 
     return qiskit_qasm3.loads(qasm3)
