+++ tests/runtime/test_quantum_job.py
@@ -15,6 +15,7 @@ Unit tests for quantum jobs functions and data types
 import pytest
 
 from qbraid.runtime.enums import JobStatus
+from qbraid.runtime.exceptions import DeviceProgramTypeMismatchError
 from qbraid.runtime.native.job import QbraidJob
 
 status_data = [
@@ -63,3 +64,37 @@ def test_map_status_raises():
     """Test raising exception for bad status type."""
     with pytest.raises(ValueError):
         QbraidJob._map_status(0)
+
+
+def get_expected_message(program, expected_type, action_type):
+    """Generate the expected error message dynamically."""
+    if program is None:
+        program_type = "NoneType"
+    else:
+        program_type = type(program).__name__
+    return (
+        f"Incompatible program type: '{program_type}'. "
+        f"Device action type '{action_type}' requires a program of type '{expected_type}'."
+    )
+
+
+class MockProgram:
+    """Mock class for testing DeviceProgramTypeMismatchError."""
+
+
+@pytest.mark.parametrize(
+    "program, expected_type, action_type",
+    [
+        (MockProgram(), "QuantumProgram", "Compute"),
+        (object(), "QuantumProgram", "Compute"),
+        (MockProgram(), "MockProgram", "Compute"),
+        (None, "QuantumProgram", "Compute"),
+    ],
+)
+def test_device_program_type_mismatch_error(program, expected_type, action_type):
+    """Test DeviceProgramTypeMismatchError with various scenarios."""
+    expected_message = get_expected_message(program, expected_type, action_type)
+    with pytest.raises(DeviceProgramTypeMismatchError) as exc_info:
+        raise DeviceProgramTypeMismatchError(program, expected_type, action_type)
+
+    assert str(exc_info.value) == expected_message
